refactor:Corrected spelling mistakes in utils.
